Refactor of recovery mode to encapsulate behavior in a class.
This introduces a WP_Recovery_Mode_Controller interface. Core would ship
with the email controller which handles recovery mode by sending an
email and setting a cookie. But this will allow for hosts or mu-plugins
to substitute any handler that implements that interface.

The cookie handler has also been abstracted out to a separate service
and can be reused by a custom handler. It also allows for substituting
all the cookie config values in the constructor. This should make it easier
to ship this eventually to network activated plugins as well, since a primary
concern there is that the correct cookie constants have not been populated
until after those plugins have been loaded.

This does not yet drop `is_protected_endpoint()` since that is still to be
discussed. Additionally, I have not had time yet to refactor the extension
storage to use one option as @flixos90 suggested.